requirements.txt
tree-sitter==0.21.3
tree-sitter-languages 
networkx 
pyyaml